Fix bug in masking logic
Previously, the masker returned the indices (in the spectrum) of
peaks that were masked as a tensor, zero-padded to the length of
the longest masked count in the batch. However, it is possible
for the 0th peak in the spectrum to be masked, which causes issues
in this setup. This change updates the logic to only consider
zeros as padding if they are not the first peak in the spectrum